feat: split css test
