+++ tests/integration/css/package.json
@@ -1,11 +1,9 @@
 {
   "private": true,
-  "name": "css-modules",
+  "name": "css-test",
   "version": "2.9.0",
   "dependencies": {
-    "react": "^18",
-    "react-dom": "^18",
-    "@modern-js/runtime": "workspace:*",
-    "@modern-js/app-tools": "workspace:*"
+    "@types/jest": "^29",
+    "@types/node": "^14"
   }
 }
